Move timestamps to int64 for easier conversion in template, moved task.id to int64 and updated cli to use int64.

+++ cmd/gogid/gogid.go
@@ -19,11 +19,11 @@ func main() {
 	printAllPtr := flag.Bool("printtasks", false, "Print entire task list to console.")
 	printDonePtr := flag.Bool("printdone", false, "Print completed tasks to console.")
 	printNotDonePtr := flag.Bool("printnotdone", false, "Print completed tasks to console.")
-	printTaskPtr := flag.Int("print", -1, "Print a single task.")
-	completeTaskPtr := flag.Int("complete", -1, "Mark a single task as completed.")
-	notCompleteTaskPtr := flag.Int("notcomplete", -1, "Mark a single task as not completed.")
+	printTaskPtr := flag.Int64("print", -1, "Print a single task.")
+	completeTaskPtr := flag.Int64("complete", -1, "Mark a single task as completed.")
+	notCompleteTaskPtr := flag.Int64("notcomplete", -1, "Mark a single task as not completed.")
 
-	idPtr := flag.Int("id", -1, "ID of task in question, used to add notes")
+	idPtr := flag.Int64("id", -1, "ID of task in question, used to add notes")
 
 	var new string
 	flag.StringVar(&new, "new", "", "Add a new task, enclose in quotes.")
@@ -67,46 +67,46 @@ func main() {
 	case "-print":
 		if *printTaskPtr > -1 {
 			taskList := fileutils.LoadTasks(tasksFile)
-			taskutils.PrintTask(&taskList, int32(*printTaskPtr))
+			taskutils.PrintTask(&taskList, *printTaskPtr)
 		}
 	case "-complete":
 		if *completeTaskPtr > -1 {
 			taskList := fileutils.LoadTasks(tasksFile)
-			taskutils.CompleteTask(&taskList, int32(*completeTaskPtr))
+			taskutils.CompleteTask(&taskList, *completeTaskPtr)
 			fileutils.WriteTasks(tasksFile, &taskList)
-			taskutils.PrintTask(&taskList, int32(*completeTaskPtr))
+			taskutils.PrintTask(&taskList, *completeTaskPtr)
 		}
 	case "-notcomplete":
 		if *notCompleteTaskPtr > -1 {
 			taskList := fileutils.LoadTasks(tasksFile)
-			taskutils.InvertCompleteTask(&taskList, int32(*notCompleteTaskPtr))
+			taskutils.InvertCompleteTask(&taskList, *notCompleteTaskPtr)
 			fileutils.WriteTasks(tasksFile, &taskList)
-			taskutils.PrintTask(&taskList, int32(*notCompleteTaskPtr))
+			taskutils.PrintTask(&taskList, *notCompleteTaskPtr)
 		}
 	case "-new":
 		if new != "" {
 			taskList := fileutils.LoadTasks(tasksFile)
 			taskutils.AddNewTask(&taskList, new)
 			fileutils.WriteTasks(tasksFile, &taskList)
-			taskutils.PrintTask(&taskList, int32(len(taskList.Task)-1))
+			taskutils.PrintTask(&taskList, int64(len(taskList.Task)-1))
 		}
 	case "-newnote":
 
 		// Requires -id, should expand command to use subcommand but this works
 		if newNote != "" && *idPtr > -1 {
 			taskList := fileutils.LoadTasks(tasksFile)
-			taskutils.AddNewNote(&taskList, int32(*idPtr), newNote)
+			taskutils.AddNewNote(&taskList, *idPtr, newNote)
 			fileutils.WriteTasks(tasksFile, &taskList)
-			taskutils.PrintTask(&taskList, int32(*idPtr))
+			taskutils.PrintTask(&taskList, *idPtr)
 		}
 	case "-nextaction":
 
 		// Requires -id, should expand command to use subcommand but this works
 		if newNextAction != "" && *idPtr > -1 {
 			taskList := fileutils.LoadTasks(tasksFile)
-			taskutils.AddNextAction(&taskList, int32(*idPtr), newNextAction)
+			taskutils.AddNextAction(&taskList, *idPtr, newNextAction)
 			fileutils.WriteTasks(tasksFile, &taskList)
-			taskutils.PrintTask(&taskList, int32(*idPtr))
+			taskutils.PrintTask(&taskList, *idPtr)
 		}
 	default:
 		flag.PrintDefaults()
