+++ model/model.go
@@ -7,10 +7,10 @@ type TaskList struct {
 
 // Task is a struct that corrisponds to a single task.
 type Task struct {
-	ID         int32    `json:"id"`
+	ID         int64    `json:"id"`
 	Task       string   `json:"task"`
-	Created    int32    `json:"created"`
-	Updated    int32    `json:"updated"`
+	Created    int64    `json:"created"`
+	Updated    int64    `json:"updated"`
 	Complete   bool     `json:"complete"`
 	NextAction string   `json:"nextaction"`
 	Notes      []string `json:"notes"`
